Fix bug where button styling is ignored
In certain cases, button styling (btn-xs, btn-info, ..) will have a lower priority in CSS than that of the generic button style, causing it to be ignored. 

Example of the issue (with/without fix): http://imgur.com/sX2uvy7

I believe that my patch addresses the issue properly, it doesn't seem to have any side effects (tested in the latest versions of Chrome, Firefox and Safari.